Add a safety buffer around detected pipes when computing path
The slight refactor that ensued was to keep the SAFETY_BUFFER
constant as private inside Driver. This meant that hitsPipe()
had to be made a (static) member. Initially, I used a lambda
for hasCrashed() to smuggle in gaps, but not it's fine to just
explicitly provide gaps to canSucceed().

I wanted to keep SAFETY_BUFFER as a member to have all the
parameters Driver cares about in one place. This has reduced
modularity somewhat (hitsPipe() is no longer a free function)
but also made its association with Driver more explicit, so win?