Streamline obstacle search
It's only necessary to find the bottom-left corner of each gap, others can be computed.
Look for the second gap even if the bird is still in the previous one - it's easy and helpful
even if we can't see the far pipe in full.
Also, look only a little ahead of the bird for the first gap - maintains the gap
the bird is in for longer.